ref(tests): Moving testing utils to testing directory
